health bars working but not aligned right

+++ 460Demo/GameCoreClasses/Orders/Abilities/AreaHeal.swift
@@ -48,7 +48,7 @@ class AreaHeal: Order, PType
             let waitAction: SKAction = SKAction.waitForDuration(wait)
             let healBlock: SKAction = SKAction.runBlock({
                 if target.alive == true {
-                    target.takeDamage(-2)
+                    target.takeDamage(-2 as Float)
                     self.healOverTime(heal-2, target:target)
                 }
             })
