+++ 460Demo/GameSceneClasses/GameScene.swift
@@ -89,7 +89,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
                         if Game.global.playerMap[playerID] != nil {
                             // if this sync msg's unitID is not my character, update it
                             if playerID != AppWarpHelper.sharedInstance.playerName {
-                                Game.global.playerMap[playerID]!.health = playerStats["health"] as Int
+                                Game.global.playerMap[playerID]!.health = playerStats["health"] as Float
                                 Game.global.playerMap[playerID]!.DS_health_txt.text = Game.global.playerMap[playerID]!.health.description
                                 
                                 let playerPos: CGPoint = CGPoint(x: (playerStats["posX"] as CGFloat), y: (playerStats["posY"] as CGFloat))
@@ -111,7 +111,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
                 if AppWarpHelper.sharedInstance.playerName != AppWarpHelper.sharedInstance.host {
                     for (enemyID, enemyStats) in syncEnemies {
                         if Game.global.enemyMap[enemyID] != nil {
-                            Game.global.enemyMap[enemyID]!.health = enemyStats["health"] as Int
+                            Game.global.enemyMap[enemyID]!.health = enemyStats["health"] as Float
                             Game.global.enemyMap[enemyID]!.DS_health_txt.text = Game.global.enemyMap[enemyID]!.health.description
                             
                             let enemyPos: CGPoint = CGPoint(x: (enemyStats["posX"] as CGFloat), y: (enemyStats["posY"] as CGFloat))
